chore: format metastore code
