+++ internal/database/metadata/postgres/entity.go
@@ -10,14 +10,6 @@ import (
 	"github.com/oom-ai/oomstore/pkg/oomstore/types"
 )
 
-func (db *DB) CreateEntity(ctx context.Context, opt types.CreateEntityOpt) error {
-	return createEntity(ctx, db, opt)
-}
-
-func (tx *Tx) CreateEntity(ctx context.Context, opt types.CreateEntityOpt) error {
-	return createEntity(ctx, tx, opt)
-}
-
 func createEntity(ctx context.Context, ext metadata.ExtContext, opt types.CreateEntityOpt) error {
 	query := "insert into feature_entity(name, length, description) values($1, $2, $3)"
 	_, err := ext.ExecContext(ctx, query, opt.Name, opt.Length, opt.Description)
@@ -29,14 +21,6 @@ func createEntity(ctx context.Context, ext metadata.ExtContext, opt types.Create
 	return err
 }
 
-func (db *DB) GetEntity(ctx context.Context, name string) (*types.Entity, error) {
-	return getEntity(ctx, db, name)
-}
-
-func (tx *Tx) GetEntity(ctx context.Context, name string) (*types.Entity, error) {
-	return getEntity(ctx, tx, name)
-}
-
 func getEntity(ctx context.Context, ext metadata.ExtContext, name string) (*types.Entity, error) {
 	var entity types.Entity
 	query := "select * from feature_entity where name = $1"
@@ -45,14 +29,6 @@ func getEntity(ctx context.Context, ext metadata.ExtContext, name string) (*type
 	}
 
 	return &entity, nil
-}
-
-func (db *DB) ListEntity(ctx context.Context) ([]*types.Entity, error) {
-	return listEntity(ctx, db)
-}
-
-func (tx *Tx) ListEntity(ctx context.Context) ([]*types.Entity, error) {
-	return listEntity(ctx, tx)
 }
 
 func listEntity(ctx context.Context, ext metadata.ExtContext) ([]*types.Entity, error) {
